Changes to handle single-precision einspline library on CPU with real builds.
To use single-precision with real builds, 
add precision="float" to <determinantset/>
  <determinantset type="bspline" href="h2o.pwscf.h5" sort="1"
                  tilematrix="1 0 0 0 1 0 0 0 1" twistnum="0" source="ion0"
                  version="0.10" precision="float">

einspline library APIs are modified 
* accept float x, float y, float z instead of double
* openmp parallelization during initialization
* initialize single-precision spline with double-precision data
* index safeguards to ensure [0,nx)x[0,ny)x[0,nz) 



git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@5621 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/einspline/multi_bspline_eval_s.h
@@ -85,25 +85,25 @@ eval_multi_UBspline_2d_s_vgh (const multi_UBspline_2d_s *spline,
 /************************************************************/
 void
 eval_multi_UBspline_3d_s (const multi_UBspline_3d_s *spline,
-			  double x, double y, double z,
+			  float x, float y, float z,
 			  float* restrict vals);
 
 void
 eval_multi_UBspline_3d_s_vg (const multi_UBspline_3d_s *spline,
-			     double x, double y, double z,
+			     float x, float y, float z,
 			     float* restrict vals,
 			     float* restrict grads);
 
 void
 eval_multi_UBspline_3d_s_vgl (const multi_UBspline_3d_s *spline,
-			      double x, double y, double z,
+			      float x, float y, float z,
 			      float* restrict vals,
 			      float* restrict grads,
 			      float* restrict lapl);
 
 void
 eval_multi_UBspline_3d_s_vgh (const multi_UBspline_3d_s *spline,
-			      double x, double y, double z,
+			      float x, float y, float z,
 			      float* restrict vals,
 			      float* restrict grads,
 			      float* restrict hess);
